feat: Adicionado metodo de validação de transação, e alteração de saldo credito ou debito ao fazer uma transação.

+++ src/main/java/com/bankWebsiteApp/demo/CardConfiguration/CardCreditLimit.java
@@ -4,7 +4,7 @@ import java.util.Random;
 
 public class CardCreditLimit {
 
-    public static int generateLimitCredit() {
+    public static double generateLimitCredit() {
         Random random = new Random();
         int minPrice = 50;
         int maxPrice = 10000;
